@@ -1,13 +1,14 @@
 {
-  "name": "react-apollo-demo",
+  "name": "datocms-yes-it-s-ananias-web",
   "version": "2.1.1",
   "private": true,
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
     "lint": "standard --fix",
     "storybook": "start-storybook -p 6006 -s public",
-    "build-storybook": "build-storybook -s public"
+    "build-storybook": "build-storybook -s public",
+    "start": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
   "eslintConfig": {
     "extends": "react-app",
@@ -39,33 +40,38 @@
     "webpack": "4.42.0"
   },
   "dependencies": {
+    "@emotion/babel-plugin": "11.11.0",
+    "@emotion/react": "11.11.3",
     "@fortawesome/fontawesome-free": "^5.15.3",
-    "@types/jest": "^27.0.3",
-    "@types/node": "^16.11.11",
-    "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
-    "apollo-cache-inmemory": "^1.6.2",
-    "apollo-client": "^2.6.3",
-    "apollo-link-http": "^1.5.15",
-    "global": "^4.4.0",
-    "graphql": "^14.4.2",
-    "graphql-tag": "^2.10.1",
+    "@vitejs/plugin-react": "4.2.1",
+    "apollo-cache-inmemory": "1.6.6",
+    "apollo-client": "2.6.10",
+    "apollo-link-http": "1.5.17",
+    "babel-plugin-styled-components": "2.1.4",
+    "browserslist-to-esbuild": "2.1.1",
+    "core-js": "3.35.1",
+    "global": "4.4.0",
+    "graphql": "16.8.1",
+    "graphql-tag": "2.12.6",
+    "lodash": "4.17.21",
     "luxon": "3.4.4",
-    "qs": "^6.7.0",
-    "react": "^16.8.6",
-    "react-apollo": "^2.5.8",
-    "react-datocms": "^1.0.5",
-    "react-dom": "^16.8.6",
-    "react-gtm-module": "^2.0.11",
-    "react-images": "^1.2.0-beta.7",
-    "react-mailchimp-subscribe": "^2.1.3",
-    "react-markdown": "^4.1.0",
-    "react-photo-gallery": "^8.0.0",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "^3.4.4",
-    "standard": "^16.0.3",
-    "styled-components": "^5.2.3",
-    "typescript": "^4.8.2"
+    "qs": "6.11.2",
+    "react": "18.2.0",
+    "react-apollo": "3.1.5",
+    "react-datocms": "4.1.3",
+    "react-dom": "18.2.0",
+    "react-gtm-module": "2.0.11",
+    "react-images": "1.2.0-beta.7",
+    "react-mailchimp-subscribe": "2.1.3",
+    "react-markdown": "9.0.1",
+    "react-photo-gallery": "8.0.0",
+    "standard": "17.1.0",
+    "styled-components": "6.1.8",
+    "typescript": "5.3.3",
+    "vite": "5.1.1",
+    "vite-plugin-babel-macros": "1.0.6",
+    "vite-tsconfig-paths": "4.3.1",
+    "wouter": "3.0.0"
   },
   "devDependencies": {
     "@storybook/addon-actions": "^6.3.12",
@@ -74,9 +80,16 @@
     "@storybook/node-logger": "^6.3.12",
     "@storybook/preset-create-react-app": "^3.2.0",
     "@storybook/react": "^6.3.12",
-    "babel-plugin-macros": "^3.1.0",
+    "@types/jest": "^27.0.3",
+    "@types/lodash": "4.14.202",
+    "@types/luxon": "3.4.2",
+    "@types/node": "18",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.0",
+    "@types/styled-components": "5.1.34",
+    "babel-plugin-macros": "3.1.0",
     "graphql.macro": "^1.4.2",
-    "sass": "true1.70.0",
+    "sass": "1.70.0",
     "storybook": "^6.3.12"
   }
 }
