@@ -11,9 +11,11 @@ import initGtm from './gtm'
 
 initGtm()
 
-ReactDOM.render(
-  <ApolloProvider client={client}>
-    <App />
-  </ApolloProvider>,
-  document.getElementById('root'),
-)
+import { createRoot } from 'react-dom/client';
+const container = document.getElementById('root');
+if (!container) throw new Error('No root element found');
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(<ApolloProvider client={client}>
+  <App />
+</ApolloProvider>,);
+
