@@ -11,7 +11,7 @@ const fragmentMatcher = new IntrospectionFragmentMatcher({
   introspectionQueryResultData: schema
 })
 
-const apiToken = process.env.REACT_APP_DATO_API_TOKEN
+const apiToken = import.meta.env.VITE_APP_DATO_API_TOKEN
 const headers = {
   Accept: 'application/json',
   'Content-Type': 'application/json',
