@@ -1,4 +1,4 @@
-import styled from 'styled-components/macro'
+import styled from 'styled-components'
 
 const TextBlockContainer = styled.div`
   max-width: 600px;
