@@ -1,5 +1,5 @@
 import React from 'react'
-import styled from 'styled-components/macro'
+import styled from 'styled-components'
 import { SpotifyEmbed } from '../molecules/SpotifyEmbed/SpotifyEmbed'
 import { FeaturedContent } from '../molecules/FeaturedContent/FeaturedContent'
 
