@@ -1,12 +1,12 @@
-import gql from 'graphql-tag'
-import React from 'react'
-import { QueryLoader } from '../../organisms/QueryLoader/QueryLoader'
-import { ReleaseCard } from '../ReleaseCard/ReleaseCard'
-import { ReleaseFragment } from '../../../queries/fragments/ReleaseFragment'
-import { VideoFragment } from '../../../queries/fragments/VideoFragment'
-import { VideoCard } from '../VideoCard/VideoCard'
-import { Video } from '../../../models/video'
-import { Release } from '../../../models/Release'
+import gql from "graphql-tag";
+import React from "react";
+import { QueryLoader } from "../../organisms/QueryLoader/QueryLoader";
+import { ReleaseCard } from "../ReleaseCard/ReleaseCard";
+import { ReleaseFragment } from "../../../queries/fragments/ReleaseFragment";
+import { VideoFragment } from "../../../queries/fragments/VideoFragment";
+import { VideoCard } from "../VideoCard/VideoCard";
+import { Video } from "../../../models/video";
+import { Release } from "../../../models/Release";
 
 const featuredContentQuery = gql`
   query FeaturedContentQuery {
@@ -21,15 +21,15 @@ const featuredContentQuery = gql`
       showFeaturedVideos
     }
   }
-`
+`;
 
 interface FeaturedContentData {
   home: {
-    featuredRelease: Release
-    showFeaturedRelease: boolean
-    featuredvideos: Video[]
-    showFeaturedVideos: boolean
-  }
+    featuredRelease: Release;
+    showFeaturedRelease: boolean;
+    featuredvideos: Video[];
+    showFeaturedVideos: boolean;
+  };
 }
 
 export const FeaturedContent = () => {
@@ -37,17 +37,18 @@ export const FeaturedContent = () => {
     <QueryLoader
       query={featuredContentQuery}
       successCallback={(data: FeaturedContentData) => {
-        const { home } = data
+        const { home } = data;
         const {
           featuredRelease,
           showFeaturedRelease,
           featuredvideos,
           showFeaturedVideos,
-        } = home
+        } = home;
         return (
           <>
             {home && showFeaturedRelease && featuredRelease ? (
               <section>
+                <h1>Featured Release</h1>
                 <div>
                   <ReleaseCard
                     release={featuredRelease}
@@ -56,18 +57,21 @@ export const FeaturedContent = () => {
                 </div>
               </section>
             ) : null}
-            {home && showFeaturedVideos && featuredvideos
-              ? featuredvideos.map((video) => (
+            {home && showFeaturedVideos && featuredvideos ? (
+              <>
+                <h1>Featured Videos</h1>{" "}
+                {featuredvideos.map((video) => (
                   <section>
                     <div>
                       <VideoCard video={video} key={video.id} />
                     </div>
                   </section>
-                ))
-              : null}
+                ))}
+              </>
+            ) : null}
           </>
-        )
+        );
       }}
     />
-  )
-}
+  );
+};
