@@ -3,9 +3,9 @@ import MailchimpSubscribe from 'react-mailchimp-subscribe'
 import { CustomForm } from './CustomForm'
 
 export const MailchimpSignupForm = (props) => {
-  const urlBase = process.env.REACT_APP_MAILCHIMP_URL
-  const mailchimp_u = process.env.REACT_APP_MAILCHIMP_U
-  const mailchimp_id = process.env.REACT_APP_MAILCHIMP_ID
+  const urlBase = import.meta.env.VITE_APP_MAILCHIMP_URL
+  const mailchimp_u = import.meta.env.VITE_APP_MAILCHIMP_U
+  const mailchimp_id = import.meta.env.VITE_APP_MAILCHIMP_ID
 
   if (!urlBase || !mailchimp_u || !mailchimp_id) {
     return <>Mailchimp Settings missing or wrongly setup.</>
