@@ -1,9 +1,11 @@
-import React from 'react'
-import gql from 'graphql-tag'
-import { Query } from 'react-apollo'
-import EventListItem from './EventListItem'
-import { LoadingPlaceholder } from '../atoms/LoadingPlaceholder/LoadingPlaceholder'
-import { ErrorMessage } from '../atoms/ErrorMessage/ErrorMessage'
+import React from "react";
+import gql from "graphql-tag";
+import { Query } from "react-apollo";
+import { EventListItem } from "./EventListItem";
+import { LoadingPlaceholder } from "../atoms/LoadingPlaceholder/LoadingPlaceholder";
+import { ErrorMessage } from "../atoms/ErrorMessage/ErrorMessage";
+import _ from "lodash";
+import { DateTime } from "luxon";
 
 const eventsQuery = gql`
   query EventsQuery {
@@ -36,48 +38,63 @@ const eventsQuery = gql`
       venueUrl
     }
   }
-`
+`;
+
+const filterUpcomingEvents = (data) => {
+  return data.allEvents.reduce((acc, event) => {
+    if (!event.date) return acc;
+    const date = new Date(event.date);
+    const today = new Date();
+
+    if (date.getFullYear() < today.getFullYear()) {
+      return acc;
+    }
+    if (date.getFullYear() === today.getFullYear()) {
+      if (date.getMonth() < today.getMonth()) return acc;
+      if (date.getMonth() === today.getMonth()) {
+        if (date.getDate() < today.getDate()) return acc;
+      }
+    }
+
+    // parse date and add to accumulator
+    acc.push({ ...event, date: DateTime.fromSQL(event.date, { zone: "utc" }) });
+    return acc;
+  }, []);
+};
+
 const UpcomingEvents = (props) => {
   return (
     <Query query={eventsQuery}>
       {({ data, loading, error }) => {
-        if (loading) return <LoadingPlaceholder />
-        if (error) return <ErrorMessage error={error} />
+        if (loading) return <LoadingPlaceholder />;
+        if (error) return <ErrorMessage error={error} />;
 
-        const upcomingEvents = data.allEvents.filter((event) => {
-          if (!event.date) return false
-          const date = new Date(event.date)
-          const today = new Date()
-
-          if (date.getFullYear() < today.getFullYear()) {
-            return false
-          }
-          if (date.getFullYear() === today.getFullYear()) {
-            if (date.getMonth() < today.getMonth()) return false
-            if (date.getMonth() === today.getMonth()) {
-              if (date.getDate() < today.getDate()) return false
-            }
-          }
-
-          return true
-        })
+        const upcomingEvents = filterUpcomingEvents(data);
+        const groupedByYear = Object.entries(
+          _.groupBy(upcomingEvents, (event) => event.date.year)
+        );
 
         return (
-          upcomingEvents &&
-          upcomingEvents.length > 0 && (
+          groupedByYear &&
+          groupedByYear.length > 0 && (
             <section>
               <h1>Upcoming Shows</h1>
               <div>
-                {upcomingEvents.map((event) => (
-                  <EventListItem key={event.id} event={event} />
+                {groupedByYear.map((eventGroup) => (
+                  <>
+                    <h2>{eventGroup[0]}</h2>
+                    {eventGroup[1].map((event) => (
+                      <EventListItem key={event.id} event={event} />
+                    ))}
+                  </>
                 ))}
               </div>
             </section>
           )
-        )
+        );
       }}
     </Query>
-  )
-}
+  );
+};
 
-export default UpcomingEvents
+export default UpcomingEvents;
