@@ -1,16 +1,12 @@
-import React, { useState, useEffect } from 'react';
-import { withRouter }  from 'react-router-dom';
-import styled from 'styled-components/macro'
-import MobileMenu from './MobileMenu'
+import React, { useEffect, useState } from 'react';
+import styled from 'styled-components';
+import MobileMenu from './MobileMenu';
 
-const BurgerButton = styled.div.attrs(props => ({
-    active: Boolean,
-  }))`
+const BurgerButton = styled.div`
   display: none;
 
   @media (max-width: 500px) {
     z-index: 1000;
-    position: ${props => props.active ? 'fixed' : 'absolute'};
     top: 16px;
     right: 16px;
     display: block;
@@ -25,7 +21,7 @@ const BurgerButton = styled.div.attrs(props => ({
   }
 `
 
-const MobileMenuButton = (props) => {
+ const MobileMenuButton = (props) => {
     const [showMenu, setShowMenu] = useState(false);
 
     useEffect(() => {
@@ -33,11 +29,11 @@ const MobileMenuButton = (props) => {
       }, [props.location]);
 
   return <>
-    <BurgerButton active={showMenu} onClick={() => setShowMenu(!showMenu)}>
+    <BurgerButton /*active={showMenu.toString()}*/ onClick={() => setShowMenu(!showMenu)}>
         {showMenu ? '╳' : '☰'}
     </BurgerButton>
     <MobileMenu visible={showMenu} />
   </>
 }
 
-export default withRouter(MobileMenuButton)
+export default MobileMenuButton
